@@ -2,9 +2,9 @@ from typing import Optional
 
 import typer
 
-from redact.v3 import InputType, JobArguments, OutputType, Region, ServiceType
 from redact.commons.utils import setup_logging
 from redact.settings import Settings
+from redact.v3 import InputType, JobArguments, OutputType, Region, ServiceType
 from redact.v3.tools.redact_file import redact_file as rdct_file
 from redact.v3.tools.redact_folder import redact_folder as rdct_folder
 
